Improved GUI keyboard controls, tooltips, and overlay robustness.
  -- Added keyboard controls to ROI viewer: space toggles mask opacity, arrows cycle color modes and colormaps.
  -- Added keyboard opacity and mask layer controls to tracking viewer.
  -- Added skip forward/backward buttons to binary viewer for frame stepping.
  -- Added tooltips to all interactive widgets across all four viewer windows.
  -- Removed keyboard shortcut references from tooltips in favor of toolbar hint labels.
  -- Added toolbar hint labels to ROI viewer and updated existing hints in other viewers.
  -- Added File button tooltips to ROI, binary, and tracking viewers.
  -- Fixed ROI overlay opacity to respect the slider value in ROIs-only background mode.
  -- Added out-of-bounds coordinate clipping for backward-deformed multi-day masks in overlays.
  -- Extracted shared QApplication setup into _get_or_create_application helper in app.py.
  -- Standardized tooltip voice to imperative tense across ROI viewer.
  -- Replaced hardcoded trace checkbox labels with TraceModeLabel enum values.

+++ src/cindra/gui/data_models.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from dataclasses import field, dataclass
 
-
 if TYPE_CHECKING:
     import numpy as np
     import pyqtgraph as pg  # type: ignore[import-untyped]
@@ -53,7 +52,7 @@ class ColorArrays:
     """The per-statistic colorbar range values as [low, mid, high] lists."""
     rgb: NDArray[np.uint8]
     """The RGBA overlay arrays with shape (color_count, height, width, 4)."""
-    random_hues: NDArray[np.float64]
+    random_hues: NDArray[np.float32]
     """The per-ROI random hue values with shape (roi_count,)."""
